requirements.txt
numpy==1.26.4
opencv-python==4.10.0.82
pillow==10.3.0
